chore: satisfy linter
